perf: speed up cache by reducing the key down to 2 git calls of work
